bingoClientDesktop/requirements.txt
flask==3.0.0
python-socketio==5.10.0
eventlet==0.33.3
